[🕹️refactor]  :  user type 분기
