+++ src/api/auth.ts
@@ -1,5 +1,6 @@
 import { supabase } from "@/supabase/supabaseClient";
-import { SignInData, SignUpData } from "@/types/components";
+import { SignInData, SignUpData } from "@/types/user";
+
 
 export const signInEmail = async (signInData: SignInData) => {
   const { data, error } = await supabase.auth.signInWithPassword({
